+++ src/main/java/com/braintreegateway/IdealPaymentDetails.java
@@ -2,6 +2,10 @@ package com.braintreegateway;
 
 import com.braintreegateway.util.NodeWrapper;
 
+// NEXT_MAJOR_VERSION Remove this class as legacy Ideal has been removed/disabled in the Braintree Gateway
+/**
+ * @deprecated If you're looking to accept iDEAL as a payment method contact us at accounts@braintreepayments.com for a solution.
+ */
 public class IdealPaymentDetails {
     private String idealPaymentId;
     private String idealTransactionId;