+++ src/main/java/com/braintreegateway/WebhookNotification.java
@@ -32,7 +32,11 @@ public class WebhookNotification {
         DISPUTE_LOST("dispute_lost"),
         DISPUTE_WON("dispute_won"),
         ACCOUNT_UPDATER_DAILY_REPORT("account_updater_daily_report"),
+        // NEXT_MAJOR_VERSION Remove this class as legacy Ideal has been removed/disabled in the Braintree Gateway
+        // DEPRECATED If you're looking to accept iDEAL as a payment method contact accounts@braintreepayments.com for a solution.
         IDEAL_PAYMENT_COMPLETE("ideal_payment_complete"),
+        // NEXT_MAJOR_VERSION Remove this class as legacy Ideal has been removed/disabled in the Braintree Gateway
+        // DEPRECATED If you're looking to accept iDEAL as a payment method contact accounts@braintreepayments.com for a solution.
         IDEAL_PAYMENT_FAILED("ideal_payment_failed"),
         // NEXT_MAJOR_VERSION remove GRANTED_PAYMENT_INSTRUMENT_UPDATE. Kind is not sent by Braintree Gateway.
         // Kind will either be GRANTOR_UPDATED_GRANTED_PAYMENT_METHOD or RECIPIENT_UPDATED_GRANTED_PAYMENT_METHOD.